Changed Date Format; DataFlow: Taskpane > Excel
We added a second taskpane in the project. A button on the 1st taskpane
navigate to the 2nd taspane that holds Radio Buttons(different date format sytes).
We needed to configure the webpack.config.js file to register the newly created
taskpane2 .
We added a 2nd button in the taskpane2.index file for navigating from
one to another taskpane.